optimize processing data points

+++ core/system_metrics.go
@@ -29,11 +29,11 @@ type CPUMetrics struct {
 
 // DiskMetrics -
 type DiskMetrics struct {
-	ReadTimeMS  int `json:"read_time_ms" bson:"read_time_ms"`
-	WriteTimeMS int `json:"write_time_ms" bson:"write_time_ms"`
-	IO          int
-	IOQueuedMS  int `json:"io_queued_ms" bson:"io_queued_ms"`
-	IOTimeMS    int `json:"io_time_ms" bson:"io_time_ms"`
-	Reads       int `json:"reads" bson:"reads"`
-	Writes      int `json:"writes" bson:"writes"`
+	ReadTimeMS  int64 `json:"read_time_ms" bson:"read_time_ms"`
+	WriteTimeMS int64 `json:"write_time_ms" bson:"write_time_ms"`
+	IO          int64
+	IOQueuedMS  int64 `json:"io_queued_ms" bson:"io_queued_ms"`
+	IOTimeMS    int64 `json:"io_time_ms" bson:"io_time_ms"`
+	Reads       int64 `json:"reads" bson:"reads"`
+	Writes      int64 `json:"writes" bson:"writes"`
 }
